Edited System Scripts
Made to be consistent with sequence_init.

Changed checks for argument counts to just be greater rather than greater than or equal.

+++ scripts/game_delete_save.gml
@@ -3,8 +3,5 @@
 
 with (ctrl_game)
 {
-    if (game_save_exists(argument0))
-    {
-        file_delete(save_directory + "save" + string(argument0) + ".sav");
-    }
+    if (game_save_exists(argument0)) file_delete(save_directory + "save" + string(argument0) + ".sav");
 }
